backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pdfplumber==0.11.4
spacy>=3.8.7,<4.0.0
sentence-transformers==3.3.1
numpy==2.2.1
pandas==2.2.3
pydantic==2.10.4
python-multipart==0.0.20
python-dotenv==1.0.1
 